LPS-72815 Moves tests
